requirements.txt
# PROJETO MCP - ANÁLISE DE VENDAS DE PEÇAS AUTOMOTIVAS
# Arquivo de dependências Python

# Conector MySQL/MariaDB
mysql-connector-python==8.1.0

# Análise de dados
pandas==2.1.3
numpy==1.25.2

# Google Gemini
google-generativeai==0.3.2

# Processamento de linguagem natural (para versões futuras)
nltk==3.8.1

# Visualização de dados (opcional)
matplotlib==3.8.2
seaborn==0.13.0

# Utilitários
python-dotenv==1.0.0
tabulate==0.9.0

# Framework web (para dashboard futuro)
flask==3.0.0

# Interface melhorada
colorama==0.4.6